requirements.txt
matplotlib
requests>=2.13.0
Pillow>=3.1.0
click>=6.7
simplejson>=3.10.0
Flask>=0.12
gunicorn>=19.6.0
itsdangerous>=0.24
Jinja2>=2.9.5
MarkupSafe>=0.23
numpy>=1.12.0
packaging>=16.8
protobuf>=3.2.0
pyparsing>=2.1.10
six>=1.10.0
tensorflow-gpu>=1.0.0
Werkzeug>=0.11.15
